pyproject.toml
[project]
name = "medical-ingestion-system"
version = "0.1.0"
description = "Agentic pipeline for ingesting and structuring medical documents using LLMs and OCR"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "google-generativeai>=0.8.6",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.10",
]
